backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-google-genai
python-dotenv
langgraph
pdfplumber
python-docx
pydantic
uuid
python-dotenv
openai
